Create mocks on slice:create
