+++ packages/manager/test/mockSlice.test.ts
@@ -1,0 +1,400 @@
+import { describe, test, expect, vi } from "vitest";
+import { WidgetTypes } from "@prismicio/types-internal/lib/customtypes/widgets";
+import {
+	SharedSlice,
+	SlicesTypes,
+} from "@prismicio/types-internal/lib/customtypes/widgets/slices";
+
+import { isRight } from "fp-ts/lib/Either";
+
+import { GeoPointContent } from "@prismicio/types-internal/lib/documents/widgets/nestable";
+import { LinkContent } from "@prismicio/types-internal/lib/documents/widgets/nestable/Link";
+import { SharedSliceContent } from "@prismicio/types-internal/lib/content";
+import {
+	DiffOperation,
+	SliceDiff,
+} from "@prismicio/types-internal/lib/customtypes/diff";
+import { mockSlice } from "../src/lib/mockSlice";
+
+vi.mock("lorem-ipsum", () => {
+	return {
+		__esModule: true,
+		LoremIpsum: vi.fn().mockImplementation(() => {
+			return {
+				generateParagraphs: vi.fn().mockReturnValue("Some text."),
+			};
+		}),
+	};
+});
+
+vi.mock("@prismicio/mocks/lib/generators/utils/slug", () => {
+	return vi.fn().mockReturnValue("Woo");
+});
+
+describe.skip("mockSlice", () => {
+	test("parse primary", () => {
+		const link = {
+			"link-2": {
+				__TYPE__: "LinkContent",
+				value: { url: "http://twitter.com", __TYPE__: "ExternalLink" },
+			},
+		};
+		const linkR = LinkContent.decode(link["link-2"].value);
+		expect(isRight(linkR)).toBeTruthy();
+
+		const geo = {
+			key: {
+				// __TYPE__: GeoPointContentType,
+				position: { lat: 48.8583736, lng: 2.2922926 },
+			},
+		};
+		const geoR = GeoPointContent.decode(geo.key);
+
+		expect(isRight(geoR)).toBeTruthy();
+	});
+
+	test("when creating a slice it should return the default mock", () => {
+		const wanted = [
+			{
+				__TYPE__: "SharedSliceContent",
+				variation: "default",
+				primary: {
+					title: {
+						__TYPE__: "StructuredTextContent",
+						value: [{ type: "heading1", content: { text: "Woo" } }],
+					},
+					description: {
+						__TYPE__: "StructuredTextContent",
+						value: [{ type: "paragraph", content: { text: "Some text." } }],
+					},
+				},
+				items: [],
+			},
+		];
+
+		const model: SharedSlice = {
+			id: "some_slice",
+			type: SlicesTypes.SharedSlice,
+			name: "SomeSlice",
+			description: "SomeSlice",
+			variations: [
+				{
+					id: "default",
+					name: "Default",
+					docURL: "...",
+					imageUrl: "",
+					version: "sktwi1xtmkfgx8626",
+					description: "SomeSlice",
+					primary: {
+						title: {
+							type: WidgetTypes.RichText,
+							config: {
+								single: "heading1",
+								label: "Title",
+								placeholder: "This is where it all begins...",
+							},
+						},
+						description: {
+							type: WidgetTypes.RichText,
+							config: {
+								single: "paragraph",
+								label: "Description",
+								placeholder: "A nice description of your product",
+							},
+						},
+					},
+				},
+			],
+		};
+
+		const result = mockSlice(model);
+
+		expect(result).toEqual(wanted);
+		// TODO: check the codec we use for SharedSliceContent[]
+		// const decoded = SliceMock.decode(result);
+		// expect(isRight(decoded)).toBeTruthy();
+		// needs to be readable by core/mocks/models SliceMock
+	});
+
+	test("when updating a mock it should append new value to existing mock", () => {
+		// "image" is missing, see below.
+		const partial: SharedSliceContent[] = [
+			{
+				__TYPE__: "SharedSliceContent",
+				variation: "default",
+				primary: {
+					title: {
+						__TYPE__: "StructuredTextContent",
+						value: [{ type: "heading1", content: { text: "Woo2" } }],
+					},
+					description: {
+						__TYPE__: "StructuredTextContent",
+						value: [{ type: "paragraph", content: { text: "Some text." } }],
+					},
+				},
+				items: [{ __TYPE__: "GroupItemContent", value: [] }],
+			},
+		];
+
+		const model: SharedSlice = {
+			id: "some_slice",
+			type: SlicesTypes.SharedSlice,
+			name: "SomeSlice",
+			description: "SomeSlice",
+			variations: [
+				{
+					id: "default",
+					name: "Default",
+					docURL: "...",
+					imageUrl: "",
+					version: "sktwi1xtmkfgx8626",
+					description: "SomeSlice",
+					primary: {
+						title: {
+							type: WidgetTypes.RichText,
+							config: {
+								single: "heading1",
+								label: "Title",
+								placeholder: "This is where it all begins...",
+							},
+						},
+						description: {
+							type: WidgetTypes.RichText,
+							config: {
+								single: "paragraph",
+								label: "Description",
+								placeholder: "A nice description of your product",
+							},
+						},
+						image: {
+							config: { label: "image", constraint: {}, thumbnails: [] },
+							type: WidgetTypes.Image,
+						},
+					},
+				},
+			],
+		};
+
+		const result = mockSlice(model, partial);
+		// text "Woo2" is still part of
+		expect(result).toMatchObject(partial);
+		// The image is random, so we check its properties instead.
+		expect(result[0].primary).toHaveProperty(
+			"image",
+			expect.objectContaining({
+				__TYPE__: "ImageContent",
+				url: expect.any(String),
+				origin: {
+					id: "main",
+					url: expect.any(String),
+					width: expect.any(Number),
+					height: expect.any(Number),
+				},
+				width: expect.any(Number),
+				height: expect.any(Number),
+				edit: { zoom: 1, crop: { x: 0, y: 0 }, background: "transparent" },
+				thumbnails: {},
+			}),
+		);
+		// TODO: check with the mock reader that this is valid
+		// const decoded = SliceMock.decode(result);
+		// expect(isRight(decoded)).toBeTruthy();
+	});
+
+	// change slice mock to the codec for SharedSliceContent[]
+	// test.skip("allFieldSliceModel", () => {
+	//   const model = Slices.toSM({ ...allFieldSliceModel });
+	//   const mock = MockSlice(model, {});
+	//
+	//   const result = SliceMock.decode(mock);
+	//   expect(isRight(result)).toBeTruthy();
+	// });
+
+	test("when i add a variation to a slice, the old mock content should be kept", () => {
+		const sliceModel: SharedSlice = {
+			id: "testing",
+			type: "SharedSlice",
+			name: "Testing",
+			description: "Testing",
+			variations: [
+				{
+					id: "default",
+					name: "Default",
+					docURL: "...",
+					version: "sktwi1xtmkfgx8626",
+					description: "Testing",
+					primary: {
+						title: {
+							type: "StructuredText",
+							config: {
+								single: "heading1",
+								label: "Title",
+								placeholder: "This is where it all begins...",
+							},
+						},
+						description: {
+							type: "StructuredText",
+							config: {
+								single: "paragraph",
+								label: "Description",
+								placeholder: "A nice description of your feature",
+							},
+						},
+					},
+					items: {},
+					imageUrl:
+						"https://images.prismic.io/slice-machine/621a5ec4-0387-4bc5-9860-2dd46cbc07cd_default_ss.png?auto=compress,format",
+				},
+				{
+					id: "foo",
+					name: "Foo",
+					docURL: "...",
+					version: "sktwi1xtmkfgx8626",
+					description: "Testing",
+					primary: {
+						title: {
+							type: "StructuredText",
+							config: {
+								single: "heading1",
+								label: "Title",
+								placeholder: "This is where it all begins...",
+							},
+						},
+						description: {
+							type: "StructuredText",
+							config: {
+								single: "paragraph",
+								label: "Description",
+								placeholder: "A nice description of your feature",
+							},
+						},
+					},
+					items: {},
+					imageUrl:
+						"https://images.prismic.io/slice-machine/621a5ec4-0387-4bc5-9860-2dd46cbc07cd_default_ss.png?auto=compress,format",
+				},
+			],
+		};
+		const previousMocks: SharedSliceContent[] = [
+			{
+				__TYPE__: "SharedSliceContent",
+				variation: "default",
+				primary: {
+					title: {
+						__TYPE__: "StructuredTextContent",
+						value: [
+							{
+								type: "heading1",
+								content: {
+									text: "Test Heading",
+									// "spans": []
+								},
+								// "direction": "ltr"
+							},
+						],
+					},
+					description: {
+						__TYPE__: "StructuredTextContent",
+						value: [
+							{
+								type: "paragraph",
+								content: {
+									text: "Some text on the default slice.",
+								},
+							},
+						],
+					},
+				},
+				items: [
+					{
+						__TYPE__: "GroupItemContent",
+						value: [],
+					},
+				],
+			},
+		];
+		const sliceDiff: SliceDiff = {
+			op: DiffOperation.Updated,
+			value: {
+				variations: {
+					foo: {
+						op: DiffOperation.Added,
+						value: {
+							id: "foo",
+							name: "Foo",
+							docURL: "...",
+							version: "sktwi1xtmkfgx8626",
+							description: "Testing",
+							primary: {
+								title: {
+									type: "StructuredText",
+									config: {
+										single: "heading1",
+										label: "Title",
+										placeholder: "This is where it all begins...",
+									},
+								},
+								description: {
+									type: "StructuredText",
+									config: {
+										single: "paragraph",
+										label: "Description",
+										placeholder: "A nice description of your feature",
+									},
+								},
+							},
+							items: {},
+							imageUrl:
+								"https://images.prismic.io/slice-machine/621a5ec4-0387-4bc5-9860-2dd46cbc07cd_default_ss.png?auto=compress,format",
+						},
+					},
+				},
+			},
+		};
+
+		const wanted = [
+			...previousMocks,
+			{
+				__TYPE__: "SharedSliceContent",
+				variation: "foo",
+				primary: {
+					title: {
+						__TYPE__: "StructuredTextContent",
+						value: [
+							{
+								type: "heading1",
+								content: {
+									text: "Woo",
+								},
+								// direction: "ltr",
+							},
+						],
+					},
+					description: {
+						__TYPE__: "StructuredTextContent",
+						value: [
+							{
+								type: "paragraph",
+								content: {
+									text: "Some text.",
+								},
+							},
+						],
+					},
+				},
+				items: [
+					{
+						__TYPE__: "GroupItemContent",
+						value: [],
+					},
+				],
+			},
+		];
+
+		const results = mockSlice(sliceModel, previousMocks, sliceDiff);
+
+		// check the content is unchanged
+		expect(results[0]).toEqual(previousMocks[0]);
+		expect(results).toEqual(wanted);
+	});
+});
