+++ packages/slice-machine/test/src/modules/selectedSlice/__fixtures__/mocks.json
@@ -1,30 +1,171 @@
 [
   {
+    "__TYPE__": "SharedSliceContent",
     "variation": "default-slice",
-    "name": "Default slice",
-    "slice_type": "dummy_slice",
-    "items": [
-      {
-        "link": {
-          "link_type": "Web",
-          "url": "http://google.com"
+    "primary": {
+      "linkLabel": {
+        "__TYPE__": "FieldContent",
+        "value": "speed",
+        "type": "Text"
+      },
+      "title": {
+        "__TYPE__": "StructuredTextContent",
+        "value": [
+          {
+            "type": "heading1",
+            "content": {
+              "text": "Higher"
+            }
+          }
+        ]
+      },
+      "link": {
+        "__TYPE__": "LinkContent",
+        "value": {
+          "id": "mock_document_id",
+          "__TYPE__": "DocumentLink"
         }
       },
-      {
-        "link": {
-          "link_type": "Web",
-          "url": "https://slicemachine.dev"
+      "imageLeft": {
+        "__TYPE__": "ImageContent",
+        "url": "https://images.unsplash.com/photo-1551739440-5dd934d3a94a?fit=crop&w=592&h=592",
+        "origin": {
+          "id": "main",
+          "url": "https://images.unsplash.com/photo-1551739440-5dd934d3a94a",
+          "width": 592,
+          "height": 592
+        },
+        "width": 592,
+        "height": 592,
+        "edit": {
+          "zoom": 1,
+          "crop": {
+            "x": 0,
+            "y": 0
+          },
+          "background": "transparent"
+        },
+        "thumbnails": {
+          "mobile": {
+            "url": "https://images.unsplash.com/photo-1515378791036-0648a3ef77b2?fit=crop&w=560&h=280",
+            "origin": {
+              "id": "mobile",
+              "url": "https://images.unsplash.com/photo-1515378791036-0648a3ef77b2",
+              "width": 560,
+              "height": 280
+            },
+            "width": 560,
+            "height": 280,
+            "edit": {
+              "zoom": 1,
+              "crop": {
+                "x": 0,
+                "y": 0
+              },
+              "background": "transparent"
+            }
+          }
         }
       },
-      {
-        "link": {
-          "link_type": "Web",
-          "url": "https://slicemachine.dev"
+      "imageLeftTitle": {
+        "__TYPE__": "FieldContent",
+        "value": "rate",
+        "type": "Text"
+      },
+      "imageLeftCta": {
+        "__TYPE__": "FieldContent",
+        "value": "funny",
+        "type": "Text"
+      },
+      "leftImageLink": {
+        "__TYPE__": "LinkContent",
+        "value": {
+          "id": "mock_document_id",
+          "__TYPE__": "DocumentLink"
+        }
+      },
+      "imageTopRight": {
+        "__TYPE__": "ImageContent",
+        "url": "https://images.unsplash.com/photo-1600861194802-a2b11076bc51?fit=crop&w=592&h=280",
+        "origin": {
+          "id": "main",
+          "url": "https://images.unsplash.com/photo-1600861194802-a2b11076bc51",
+          "width": 592,
+          "height": 280
+        },
+        "width": 592,
+        "height": 280,
+        "edit": {
+          "zoom": 1,
+          "crop": {
+            "x": 0,
+            "y": 0
+          },
+          "background": "transparent"
+        },
+        "thumbnails": {}
+      },
+      "imageTopRightTitle": {
+        "__TYPE__": "FieldContent",
+        "value": "soil",
+        "type": "Text"
+      },
+      "imageTopRightCta": {
+        "__TYPE__": "FieldContent",
+        "value": "printed",
+        "type": "Text"
+      },
+      "topRightImageLink": {
+        "__TYPE__": "LinkContent",
+        "value": {
+          "id": "mock_document_id",
+          "__TYPE__": "DocumentLink"
+        }
+      },
+      "imageBottomRight": {
+        "__TYPE__": "ImageContent",
+        "url": "https://images.unsplash.com/photo-1607582278043-57198ac8da43?fit=crop&w=592&h=280",
+        "origin": {
+          "id": "main",
+          "url": "https://images.unsplash.com/photo-1607582278043-57198ac8da43",
+          "width": 592,
+          "height": 280
+        },
+        "width": 592,
+        "height": 280,
+        "edit": {
+          "zoom": 1,
+          "crop": {
+            "x": 0,
+            "y": 0
+          },
+          "background": "transparent"
+        },
+        "thumbnails": {}
+      },
+      "imageBottomRightTitle": {
+        "__TYPE__": "FieldContent",
+        "value": "smile",
+        "type": "Text"
+      },
+      "imageBottomRightCta": {
+        "__TYPE__": "FieldContent",
+        "value": "exchange",
+        "type": "Text"
+      },
+      "bottomRightImageLink": {
+        "__TYPE__": "LinkContent",
+        "value": {
+          "id": "mock_document_id",
+          "__TYPE__": "DocumentLink"
         }
       }
-    ],
-    "primary": {
-      "section_title": "enhance real-time architectures"
-    }
+    },
+    "items": [
+      {
+        "__TYPE__": "GroupItemContent",
+        "value": []
+      }
+    ]
   }
 ]
